Deprecate libavb_ab and bootctrl.avb code.
This code was already marked as experimental in anticipation of being
removed in the future. Officially deprecate it and set Jun 1 2018 as
the date it will be removed. This should give users of the code time
ample time to fork/migrate.

To keep using the code AVB_AB_I_UNDERSTAND_LIBAVB_AB_IS_DEPRECATED
must be defined.

The reason for deprecating this code is twofold:

 - Its policy was optimized for devices without a display with
   e.g. automatic fallback to the other slot if a slot fails to
   boot. Since most A/B stacks in Android devices don't work this
   way this code is confusing.

 - There are no known active users, no good test coverage for the
   bootctrl.avb code, and no plans to use it.

When the code is removed we'll provide an easy transition path by
keeping (but renaming) the |ab_ops| member in AvbOps.

Bug: 64140612
Test: All unit tests pass.
Test: UEFI-based boot loader compiles.
Change-Id: Id5e090a2048076d36ccca2e1c4cb55e226b8b43d