Update patch set 7
Patch Set 7:

> Patch Set 7: Code-Review-1
> 
> Uh, I don't think this is a good idea. Let me try to understand what you are trying to achieve...
> 
> Are you trying to, say, put e.g. avb_crypto.h in one directory and avb_crypto.c in another? If so, please don't do that... the intent is that you simply copy libavb/*.[ch] into your project and build a library out of what's in libavb/*. Specifically you should never need to modify anything in your local libavb/ directory... this is so it's easy to upgrade to a newer libavb version.

but it is not how uboot (or linux) organizes its source code. having them in the same folder makes the libavb look very different in their source code tree. Also, to import and use the libavb headers in other files (like in a uboot command implementation), I have to add custom "-I" flags pointing to the folder just for referencing libavb headers.

I think the reasoning of uboot and linux is that a C library/package/module needs its public headers exported to be usable, so these public headers should be separated from the .c files that do not need to be exported (e.g. for linux plugable modules). This makes it easy to export a library in its binary form.

I am just hoping that libavb can blend in with some consistency.

The change also does not hurt libavb very much. To practically have all these extensions working properly together, the include path needs to be pointed to the shared root of these libavb* folders. The only benefit I see of using a relative import path is that the sole "libavb" directory can then be put at any location and built alone without specifying a special import path.

practically, it is unlikely for libavb upgrade to be simple full directory copy-in. for example, in uboot, if a file is added or removed, I have to update the Makefile in each directory either manually or via a script. Loon uses chromeos as its base system, which uses gentoo's emerge build system. our current way is to copy the source files into the ebuild package directory, and then use the scripts to copy all the .h files into the "include/libavb" folder, and ".c" files into the "libavb" source directory.

Patch-set: 7
Reviewer: Lonnie Liu <1242600@85c56323-6fa9-3386-8a01-6480fb634889>