Update patch set 1
Patch Set 1: Code-Review-2

I don't think this is right ... if verification is disabled we should do as little as possible to get the device booting as if AVB wasn't implemented in the bootloader. This behavior is already documented here: https://android.googlesource.com/platform/external/avb/+/master/libavb/avb_slot_verify.c#1231

If code in the HLOS is expecting androidboot.veritymode to be 'disabled' that's a problem with that code. It should be fixed to expect either 'disabled' or it to not to be set.

Patch-set: 1
Reviewer: David Zeuthen <1076971@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2