libavb: Support storing vbmeta blobs in dedicated partitions.
With this commit, libavb will look in the partition with the name
vbmeta_${PART_NAME}${AB_SUFFIX} for a vbmeta blob before looking for a
footer in the ${PART_NAME}${AB_SUFFIX} partition. This is useful in
situations with non-standard partitioning setups.

Previously it was optional to support the get_size_of_partition()
operation. Starting with this commit, this operation is mandatory and
must be implemented such that AVB_IO_RESULT_ERROR_NO_SUCH_PARTITION is
returned for partitions that don't exist.

Also introduce a new 'extract_vbmeta_image' sub-command to avbtool
that can be used to extract the vbmeta blob from a footer.

Bug: 80195851
Test: New unit tests and all unit tests pass.
Change-Id: I92190259ff5c62894fc54081ff9e8bb5205e2786