Update patch set 1
Patch Set 1:

(3 comments)

Thanks for looking.

> Patch Set 1: Code-Review+1
> 
> (3 comments)
> 
> LGTM, let me adding Sandeep to confirm the 1:1 or 1-to-many things.
> 
> Btw, do we need to bump minor version? 
> Because we're going to mandate get_size_of_partition(), which might not be implemented in legacy avb clients.
> 
> https://android.googlesource.com/platform/external/avb/#Versioning-and-Compatibility

We do not - that section is only about the file format. Mandating get_size_of_partition() is about the API and we've already said the following about API (in the "Portability" section): "The intention is to keep the API of the library stable however it will be broken if necessary." ... we've previously broken API, it's fine. It just means the integrator need to adapt for that slight change when they integrate a new version of libavb.

Patch-set: 1
Reviewer: Gerrit User 1076971 <1076971@85c56323-6fa9-3386-8a01-6480fb634889>