libavb: Use size_t instead of uint32_t for lengths in avb_sha*() functions.
We won't run into this problem with the way we use SHA-256 or SHA-512
(the partitions with hash descriptors - e.g. boot.img and others - are
typically less than 4 GiB) but it might be nice to fix anyway.  Bowgo
pointed this out a while ago but we never fixed it.

Also add a new unit test to fix this.

Bug: None
Test: New unit tests + all unit tests still pass (both on 32- and 64-bit).
Change-Id: I964ee1056c61c4591ff4c1d93ad8a1d1676e2632