Only automatically initialize new persistent digest values when locked
This is a modification of the change made in go/oag/706769, which added
automatic initialization of persistent digest values when no value
exists, to only do this initialization when the device is locked.
Without this change, it is likely that - in a typical factory flow - the
initialization will occur before the contents of the partition have
actually been populated. This would require the factory to manually
clear the persistent digest value before locking, which is an undesired
extra step.

The error handling in read_persistent_value has also been adjusted so
that the write errors are handled explicitly, rather than falling
through to the read error handling. This avoids getting a secondary and
confusing error message, e.g. "Digest does not exist" -> "Error
initializing" -> "Error reading", if initialization fails. The unit
tests are also updated to test the read and write errors separately.

Bug: 117298653
Test: libavb_host_unittest
Change-Id: I448a9b7684cc2852ff5b92359a3d6e2931aee8ba