at_auth_unlock.py: Choose matching unlock credentials if given multiple
This adds support for passing in multiple unlock credential archives
and/or directories containing such credential archives. The tool will
automatically search all provided archives and automatically use any
the first match found. Matches are found based on the SHA256 hash of the
product ID, which is included in the unlock certificate (in the archive)
and in the challenge from the device.

Fixes: 118340336
Test: './at_auth_unlock.py .' with no/one/multiple archives in PWD,
verified that tool finds the match and uses it if a match exists.
Change-Id: If4980d535039b3da0b0f08482cf2a22ccf8bbfb6