Fix a bug that would cause OoB memory read.
The magic number in avb_vbmeta_image_verify is checked before the
length is checked, thus if the structure size is less than the magic
number length, OoB memory would be read. This bug is found by running
a fuzz test that gives empty input to this function.

Tests: Added unit test.
Bug: b/133164384