Update patch set 6
Patch Set 6:

> Patch Set 6: Autosubmit+1
> 
> (1 comment)
> 
> #@$#@! FFS, really annoying how TreeHugger insists on that all comments should be resolved.

Yup, agree.
It's really annoying and sometimes it's hard to figure out which comments are not resolved yet. (especially there are lots of patch-sets...).

I think the original intention is good, but gerrit probably needs to have some method for us to find out those comments easily.

Thanks David for resolving this with autosubmit!

Patch-set: 6