Create change
Uploaded patch set 1.

Patch-set: 1
Change-id: Ib5242d536f6c7c3d976eec21dbe8d64304d23922
Subject: This change is part of a topic that moves the recovery resources from the system partition to the vendor partition, if it exists, or the vendor directory on the system partition otherwise. The recovery resources are moving from the system image to the vendor partition so that a single system image may be used with either an A/B or a non-A/B vendor image. The topic removes a delta in the system image that prevented such reuse in the past.
Branch: refs/heads/master
Status: new
Topic: 
Commit: dab01c36e6c4edb1750294733c6637a542ae504a
Tag: autogenerated:gerrit:newPatchSet
Groups: dab01c36e6c4edb1750294733c6637a542ae504a
Private: false
Work-in-progress: false