Update patch set 4
Patch Set 4:

> The plan was originally to store it in a separate partition, but after
> discussions, we decided that the additional overhead for OEMs to generate
> a new partition, especially a smaller one, was not optimal.

True, new partitions also make upgrades harder. And it has to be a physical partition since we can't rely on dynamic partitions / super this early on in the boot process.

> It was
> suggested to us to store it at the end of the vbmeta partition in a way
> that would not interfere with the integrity of the vbmeta information. 

It's definitely possible to do that and I think it should be OK... just please make sure the AFTL data has magic + size so it's easy for other code to identify and skip the AFTL data in case we ever want to put other stuff after vbmeta + AFTL data.

FWIW, I think the build system already pads vbmeta.img with zeroes to make its size a multiple of 4 KiB (this was added because some fastboot implementations choked otherwise).

Btw, your code for loading this data for evaluating the proof at runtime [0] may need to have the same quirks/workarounds as avb_slot_verify() though (as will your code for appending it to the end of the vbmeta struct)... from the top of my head this currently only includes looking for 'vbmeta' at the end of 'boot' if there is no 'vbmeta' partition [1]. Why do we support this? IIRC it had to do with upgrade-path for existing devices - they didn't want to add a 'vbmeta' partition (sounds familiar? :-) ...). It might be easiest to say that AFTL just don't support systems without a vbmeta partition... that way you won't have to add a lot of code to build and verification paths which is annoying to test etc.

[0] : should probably go in libavb_aftl/ similar to how we have libavb_atx/ ... can still use all the platform abstractions we have in libavb (e.g. AvbOps, avb

[1] : https://android.googlesource.com/platform/external/avb/+/master/libavb/avb_slot_verify.c#704

> Will update the readme to include AFTL affected flow. :)

Excellent, thanks!

Patch-set: 4
Reviewer: Gerrit User 1076971 <1076971@85c56323-6fa9-3386-8a01-6480fb634889>