avbtool: avoid repeated string concatentations
Python strings are immutable, so concatenating every hash for a
level onto a string causes the string to be repeatedly reallocated.
For a SANITIZE_HOST=address build this caused avbtool to take 90
minutes to sign a system image.  Append to a list instead, and
concatentate the list at the end of the level.

Bug: 140533751
Bug: 150246447
Test: m systemimage
Change-Id: Idbbdb075d860d7e834e3d130b1968ac380eace31