Pack AftlIcpEntry and fix parse_icp_entry
Update AftlIcpEntry definition to be packed. Without it, the
inc_proof_size field of tmp_icp_entry is not guarantee to be at a
constant offset.

Update the function definition and body to correctly advance aftl_blob
while reading it. Update the unit tests with new data (based on aftltool
unit tests) and drop exact value comparisons.

Test: m && ./out/host/linux-x86/nativetest/libavb_host_unittest/libavb_host_unittest
Test: m && ./out/host/linux-x86/nativetest64/libavb_host_unittest/libavb_host_unittest
Fix: 150876501
Bug: 150353568
Change-Id: I1d6043446a900684b0e261b7ab12e61bc0ea1158