Update patch set 1
Patch Set 1:

If we're adding avbtool3 (as I suggested), we also need to remove avbtool. Plus deal with all the fallout from this (update AOSP build system, etc.). The latter would be a lot of work.

Maybe the easiest thing is just to keep the name avbtool but just update the hashbang from

 #!/usr/bin/env python

to

 #!/usr/bin/env python3

I think this might actually work. Let me ask around about this...

Patch-set: 1
CC: Gerrit User 1076971 <1076971@85c56323-6fa9-3386-8a01-6480fb634889>