Fixing up interactions with subprocess.Popen to be py3 compatible.
Going forward with Python 3 all interactions with subprocess are with
byte-like objects unless an encoding was specified when calling Popen()
and return values are of type bytes().

Going forward I will standardize on using bytes() instead of bytearray()
except for situations where the mutable propertly of bytesarray() is
actually used.

This change is still compatible with Python 2.

Bug: 151336743
Test: atest --host libavb_host_unittest
Test: atest --host aftltool_test
Test: ./aftltool_integration_test.py
Change-Id: I8d6c5459ce56ade8fee1e258a5309a261188ff21