Merge transparency log arguments
Merge the transparency_log_servers and transparency_log_pub_keys
arguments into one. This is to allow for further attribute definitions
per log. For instance, to use an api_key (implemented but not used).

For make_icp_from_vbmeta subcommand, the argument is for example:
--transparency_log_servers example.com:8080,key1.pub \
  example.com:8081,key2.pub

Introduce a TransparencyLog class to group the data. It is passed
down the call chain instead of the host and public keys as separate
arguments.

Test: atest --host aftltool_test
Bug: 153940575
Change-Id: I5ad8a5f5ef8eb131e496e2d6a99a77ce4d104139