Final switch from Python 2 to Python 3 for avbtool.
As of January 1st, 2020 Python 2 is not officially supported anymore:
https://www.python.org/dev/peps/pep-0373/#id4

As a consequence avbtool has been ported to Python 3 over the last
couple of changes in a still Python 2 compatible way. This particular
change is the last step in the porting work to switch over to Python 3
and breaks backward compatibility with Python 2 runtime environments.

Removed the avbtool_test.py since libavb_host_unittest already provided
sufficient coverage for supporting the porting work.

Replaced the binascii.hexlify(x).decode('ascii') construct, which worked
for both py2/py3, to x.hex(), which now only works on py3 but is better
readable.

Bug: 151336743
Bug: 153057071
Test: atest --host libavb_host_unittest
Test: atest --host aftltool_test
Test: ./aftltool_integration_test.py
Test: ./aftltool info_image_icp --vbmeta_image_path test/data/aftltool/aftl_output_vbmeta_with_2_icp_same_log.img
Test: ./aftltool info_image_icp --vbmeta_image_path test/data/aftltool/aftl_output_vbmeta_with_2_icp_different_logs.img
Change-Id: I66fb8afc2c62451e47b81509b2ed0197246e7737