Final switch from Python 2 to Python 3 for avbtool.
As of January 1st, 2020 Python 2 is not officially supported anymore:
https://www.python.org/dev/peps/pep-0373/#id4

As a consequence avbtool has been ported to Python 3 over the last
couple of changes in a still Python 2 compatible way. This particular
change is the last step in the porting work to switch over to Python 3
and breaks backward compatibility with Python 2 runtime environments.

Bug: 151336743
Test: atest --host libavb_host_unittest
Test: atest --host aftltool_test
Test: ./aftltool_integration_test.py
Change-Id: I0e8c31112d238cb2aa3f0a2e90d3b9097b8c78be