Change log leaf encoding
Previously, the log leaves were JSON-encoded, making the parsing
over complicated for the bootloader. Move to TLS-presentation encoding
which can deterministacly be reproduced.

Use the BytesIO type when encoding/decoding to avoid the repeating
pattern of reading a value and advancing in the buffer.

Drop the 2 ICPs from differrent logs tests.

Test: aftltool_test.py
Test: aftltool_integration_test.py
Bug: 154303414
Change-Id: I565c478bc5b8cef4c6fdde8f558760334daf4b55