Update patch set 5
Patch Set 5:

Sorry for the delay, finally had some time to look at it. The proposed change is wrong and I'll explain why. It's a bit subtle so bear with me.

You will note that the code you modified has the following comment

  /* Update stored rollback index such that the stored rollback index
   * is the largest value supporting all currently bootable slots. Do
   * this for every rollback index location.
   */

The key part here is this: "[...] largest value supporting all currently bootable slots [...]"

Essentially it means that we cannot pick the biggest rollback index of _all_ slots... no, we have to actually pick the smallest one... because what happens if you just pick the biggest one like you suggest? Well, you then make the slot with that smallest rollback index unbootable. And this is what we cannot do.

If you look at the test case you had to modify, this is exactly what it tests. The preconditions are

 slot A: rollback indexes 4 and 9
 slot B: rollback indexes 5 and 7

after

Patch-set: 5
Reviewer: Gerrit User 1076971 <1076971@85c56323-6fa9-3386-8a01-6480fb634889>