Update patch set 5
Patch Set 5: Code-Review-2

Sorry I hit ctrl+enter or something which caused it to be sent prematurely. Here's the rest of my comment:

If you look at the test case you had to modify, this is exactly what it tests. The preconditions are for rollback indexes on partitions 'boot' and 'odm' are as follows:

 slot A: rollback indexes 4 and 9
 slot B: rollback indexes 5 and 7

and the post-condition we want is indeed 4 (smallest rollback index for partition 'boot') and 7 (smallest rollback index for partition 'odm') as the test checks for. Your change would make this 5 and 9 and with this change you now can't boot either slot A or slot B.. which would result in a bricked device.

Hope this all makes sense. If not, please do ask.

Patch-set: 5
Reviewer: Gerrit User 1076971 <1076971@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2