Extension to link AVB with Open Profile for DICE
After running avb_slot_verify(), use the verified data as inputs to the
Open Profile for DICE. The certificates conform to Android's BCC format.

The AVB digest is used as the code hash input and a hash of the public
key of the top-level VBMeta struct is included in the config input.

Test: --host libavb_host_unittest
Change-Id: Ie4605ba8cd87246e457c94345000d50bec856032