libavb: clarifying hashtree_error_mode usage
The hashtree error mode can be categorized into three groups:

* AVB_HASHTREE_ERROR_MODE_RESTART_AND_INVALIDATE, which needs
    CONFIG_DM_VERITY_AVB in the kernel to invalidate the current slot.

* AVB_HASHTREE_ERROR_MODE_RESTART and AVB_HASHTREE_ERROR_MODE_EIO

* AVB_HASHTREE_ERROR_MODE_MANAGED_RESTART_AND_EIO

The first method is for legacy usage and is not recommended.
Because after falling back to the other slot, the the device might not
be able to read /data.

The second and the third methods are similar. The difference is the
owner who switches between EIO and RESTART modes.

Bug: None
Test: presubmit

Change-Id: If9ff9d68cd4a3d316a613829797bc746932f7c30