tools/transparency: Fix encoding bugs and add log payload.
Bug: 202855583

Fixed an issue with string operation due to treating output from
stdout pipe as `str`, when it is actually a `bytes` object. This is fixed by
specifying encoding as UTF-8 when doing Popen().

Additionally, added code to correctly format contents for log payload
and write to a file named log_payload.txt upon success. User should be
able to use that file as an input to perform log inclusion test.

Change-Id: Idcd04b2b68d1b8e7f0232c13e0c466250cd11a8e