libavb: Keep types in the SHA context
The context used to reserve a buffer sized for the implementation's
context, however there was an unnecessary loss of typing. Change the
interface to expect a type for the implementation's context rather than
simply a size. The implementation could use an opaque buffer, should it
choose, but can also use types directly, especially if it was already
using sizeof to define the size.

Test: (cd external/avb && atest --host-unit-test-only)
Bug: 185329132
Change-Id: I7e483479ebfac56f0264b287daeea61d391aec37