avbtool: support add_hashtree_footer for very small image file
Currently the following command will emit unclear error message:
  1. `dd if=/dev/urandom of=4k.img bs=1k count=4`
  2. `avbtool add_hashtree_footer --image 4k.img --algorithm
      SHA256_RSA2048 --key ./external/avb/test/data/testkey_rsa2048.pem`

Error message:
  out/host/linux-x86/bin/avbtool: Adding hashtree_footer failed: local
  variable 'level_output' referenced before assignment.

The problem is that generate_hash_tree() doesn't consider the
case that the image size is just one block. Fixing this by
returning the hash of the image as the root hash with tree_size
as zero, if the image size is just one block.

Note: the image size must be a multiple of block size.

Bug: 210696040
Test: libavb_host_unittest

Change-Id: I91d689ed368187762bc7e6c17332cc6bbba01b23