Fix test build rules for different SHA implementations
The tests were linking against libraries that had different expectations
for the SHA implementation leading to memory issues. Resolve the
conflicts but building the dependencies from source so the correct SHA
implementation is used by all the code under test.

The problems was, in part, hidden due to an atest bug b/211900473 that
meant the custom SHA implementation variant of the test wasn't actually
being run. It needs to be run manually to confirm the fix until that is
resolved.

Test: ./libavb_host_unittest_sha
Bug: 211287350
Change-Id: I32b944fde9c133d1b9697a509ff66a26e084d6c4