[avb] Test partition before allocating 64KB memory for VBMeta
This cl is an optimization for load_and_verify_vbmeta(). When
this method is invoked to load VBMeta, prior to this cl, we
allocated 64KB memory and then tries to read the partition
`vbmeta`. This cl added a test before this action, so we first
try to read 1 byte from the end, we only allocate th 64KB memory
when it is successful to read 1 byte.

This unblocks the cl aosp/2388176 where we don't have the
`vbmeta` partition and also cannot afford allocating the 64KB
memory in pVM firmware.

Bug: 256148034
Test: atest libavb_host_unittest libpvmfw_avb.integration_test
Change-Id: Ia242ddedc8bb5e8a4277a9a766d563d8ae8cdd60