avb: add Rust bindgen packages
Moves the Rust bindgen rules from //packages/modules/Virtualization into
libavb so they can live alongside the C implementation.

For now the bindgen is exposed to Virtualization directly so it
continues to work, but future CLs will move the higher-level safe Rust
wrapper here as well so that Virtualization and other packages can
depend on it instead, as recommended by Android bindgen best-practices.

Bug: 290110273
Test: atest
Change-Id: Id70b88b5021d1b224bd7a8b23ed953fab6a926d8