libavb_rs: fix default GUID callback behavior
The default GUID callback (when the user doesn't want this feature)
turned out to be incorrect for verification. Whenever a separate vbmeta
partition is used, libavb will automatically attempt to inject its GUID
into the kernel commandline, and erroring out from this callback results
in verification failure.

Instead, for the vbmeta* partitions specifically, we need this callback
to succeed, but the actual GUID returned doesn't matter since nobody is
using it so we give the empty string.

Bug: b/290110273
Bug: b/308829443
Test: atest
Change-Id: Ied33940be8bd35eb9da536dd3d40bbb1c65a4e0d