libavb_rs: add filter_descriptors()
Adds a utility function filter_descriptors() which allows callers to
replace something like this:

```
descriptors
    .iter()
    .filter_map(|d| match d {
        Descriptor::Hash(h) => Some(h),
        _ => None,
    })
```

with this:

```
filter_descriptors::<HashDescriptor>(descriptors);
```

This takes a surprising amount of code and some decently complicated
lifetime specifications, but this is such a common need for users that
it seems worth providing here.

If we don't provide this, it also seems likely that users may attempt to
implement it themselves as a local helper, and end up running into all
this potentially frustrating complexity anyway.

Bug: b/320542357
Test: atest libavb_rs_test libavb_rs_uuid_test libavb_rs_unittest libavb_rs_uuid_unittest
Change-Id: I20622a36aaa3afbc857826da883369ed6cd9c403