libavb_rs: make lifetimes more flexible
Previously we required that the Ops object and the preloaded partition
data it wrapped had the same lifetime. This created a bit of awkwardness
because the preloaded data gets passed into the verification result,
which means the verification result also carried an (unnecessary) borrow
of the entire Ops object.

This CL splits the preloaded data into its own lifetime. This allows the
verification result to not care about the Ops object, and only borrow
the preloaded data, giving the caller much more flexibility.

Bug: b/330335215
Test: atest libavb_rs_test libavb_rs_uuid_test libavb_rs_unittest libavb_rs_uuid_unittest
Test: `mm` with the compile-fail test included fails to build as expected
Change-Id: If5a1f3f0ca7235532b511143e39abf4b551fc0f9