libavb_rs: ops framework for ATX
Adds the framework for the ATX operations. Nothing is implemented yet,
just the necessary refactoring to support the additional data.

The restructuring should hopefully be a bit cleaner and more readable
overall; all the data needed for the callbacks is now gathered in the
`UserData` class, and there's some additional compile-time protection
against pointer invalidation by encapsulating the raw pointers more
tightly.

Bug: b/320543206
Test: atest libavb_rs_test libavb_rs_uuid_test libavb_rs_unittest libavb_rs_uuid_unittest
Change-Id: Ice96f93795dbe76cdd5f06c542172b95a253cdd9