libavb: rename "ATX" to "cert"
ATX (Android Things eXtensions) is a misleading name because it is still
useful outside the deprecated Android Things project. It isn't tied to
anything AT-specific but is instead a general certificate-based
authorization mechanism that could be valuable to many AVB users.

To better represent this, rename "ATX" to "cert" to better describe its
purpose and functionality.

No functional changes have been made here, it's purely renaming and
reformatting.

Bug: b/333078493
Test: m
Change-Id: I5d2f981a894db92dbd268cb0fcb53c21edf4e986