libavb: refactor cert usage strings
Refactors the certificate usage strings:
 * pulls them into constants
 * adds a bit of documentation
 * adds a `--usage_for_unlock` arg to `avbtool make_certificate`

The intent here is to avoid users having to deal with the usage strings
directly, and to provide some context for why they contain the
substrings "android.things" even though they aren't specific to the
Android Things project.

In addition, the usage hash values are now pre-calculated in the
libavb_cert implementation. This isn't directly related to the
"android.things" string, but seems worth doing while we're here:
 * reduces .a artifact sizes by 1-2KiB
 * should reduce runtime stack usage by 32 bytes for the cert APIs

Bug: b/333078493
Test: atest libavb_host_unittest
Change-Id: I041b7f1b45e14657da1c3897085bdcc249498f52