Remove usage of libchrome's base::StringPrintf
As part of the effort to uprev libchrome in Android, we are
removing Android dependencies to libchrome utility functions
that are not strictly required.

This change removes usages of base::StringPrintf in avb tests,
and substitutes them with the equivalent function in Android's
libbase.

Bug: 360917504
Test: atest libavb_host_unittest
Change-Id: I1882e07f662f9f8ce0c2bf990ac94aa3a30427fa
Merged-In: I1882e07f662f9f8ce0c2bf990ac94aa3a30427fa