[1/N] Remove usage of libchrome's base::FilePath from avb tests
As part of the effort to uprev libchrome in Android, we are
removing Android dependencies to libchrome utility functions
that are not strictly required.

This change substitute many base::FilePath from avb tests
with std::filesystem::path. In the process, it also transitions
many usages of base::ReadFileToString to
android::base::ReadFileToString.

Bug: 360917504
Test: atest libavb_host_unittest

Change-Id: Id3daf7dfe0d72753d2bc014dfef167354c92b359