tools/transparency/verify: Refactor verifier to handle new log.
Bug: 370743686

A new log was introduced at
https://developers.google.com/android/binary_transparency/google1p/checkpoint.txt
and
https://developers.google.com/android/binary_transparency/google1p/checkpoint.txt.
This change parameterized previously hardcoded addresses and variables
so that users can also use this verifier to perform inclusion proof test
on both logs.

Also added a new public key that corresponds to the key that used to
sign the new log. This is so that users do not have to fetch the keys
manually.

Test: Manual.
Change-Id: Id3b5047a000fa0d9aee93f72c90690862f8e4a0c