Change avbtool to also check for files in uppercase instead only using the partition name. For example some firmware images contain a VBMETA.img with all different partitions such as system, vendor, etc., however the files are also in all uppercase for example SYSTEM.img. These files are not found in the current version of the tool.
Change behavior of verify_image when a partition is not found; the current version raises an AvbError and stops the verification completely.
Users of the avbtool might still be interested in the verification of the remaining partitions. This might be a valid use case with incremental OTA updates.
The proposed changes still raise an AvbError, but only after the whole partition chain is verified.

Add new argument --allow_missing_partitions to the verify_image sub_parser. This argument prevents ImageHandler to throw a FileNotFoundError when the partition's .img file is not found.
The default value of --allow_missing_partitions is False.
By default the verification fails if a) one of the partitions could not be verified or b) a partition is missing.
--allow_missing_partitions only skips the b) part of the checks.

Change-Id: I8d44b5a388f8e8f8c5549a509c1dbf73af03d63a