Revert "Adding hash partitions digest and hash algorithm in ro p..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/60445089-769a-446c-94f6-85cd86d31114). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8590c9c57b52d33f22bb1e1d18142b5f82dd1fee
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com