Revert "Adding hash partitions digest and hash algorithm in ro p..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b19cb3cf-d050-4968-9f7a-1ea72ccdd27a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia9c27f83f28e646144be99d4bc72de9ac28368d8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com