Revert "Adding hash partitions digest and hash algorithm in ro p..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7da1211e-35de-4d16-8388-afb8b22ecfd7). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id06dc0b5166016beb38528904f9525d3d4c8efdb
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com