Revert "Adding hash partitions digest and hash algorithm in ro p..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b8efa414-87ad-4b9a-8f65-0899cf199269). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib5b273df02548832d2cc0e305866bce504f69cb5
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com