Revert "Adding hash partitions digest and hash algorithm in ro p..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d2723707-efd0-465b-acc1-fda900a62bbc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: If118b37ec4d570cda574ad8bfeee4034b5965bfa
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com