Revert "Adding hash partitions digest and hash algorithm in ro p..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/855250cd-80b7-4f86-b65b-34e7ed840f1d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5e5607a8c8186da72934d87e990100de3eb269a9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com