Revert "Adding hash partitions digest and hash algorithm in ro p..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7d772806-af2d-467a-9677-05bf6540d559). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I5a09ab6279f8385590030f28fc40fab00d371961
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com