Sort chained partitions
The order of the chained partitions shouldn't matter, but it does
affect the output file. Sort the partitions so that you get a consistent
result across different usages of avbtools that may pass the flags
in a different order. My motivation is for comparing soong+make builds
to soong-only builds.

Bug: 390204058
Test: Presubmits
Change-Id: Ic2d728dd65a05a482f9fcddd8095651172a2c783