avbtool.py: Add "update_partition_descriptor" subcommand
See the Avb.update_partition_descriptor() method's docstring for
motivation.

This subcommand's implementation is heavily based on
Avb.make_vbmeta_image(). I recommend referring to said method during
review.

The unit tests in this CL ensure that this command only modifies the
expected hash or hashtree descriptor in the input image while leaving
everything else intact.

I also tested this subcommand by flashing the "boot" partition on my
device with a different kernel, as well as the "vbmeta" partition with
the result of patching the device's original vbmeta image with this
subcommand, and verifying that 1) the device boots in developer mode as
well as normal mode, and 2) there are no AVB-related error messages in
the kernel and firmware logs. The specific invocation I used for this
test is as follows:

    ./avbtool.py update_partition_descriptor \
        --image vbmeta.img \
        --partition_image boot.img \
        --output vbmeta-new.img \
        --algorithm SHA256_RSA4096 \
        --key path/to/key.pem

Bug: 433986426
Bug: 408522717
Test: m libavb_host_unittest && libavb_host_unittest
Change-Id: I253444fe3f5c1ac6097cb76350a84fa41a6d8398